+++ src/main/java/cart/domain/cartItem/CartItem.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.cartItem;
 
 import cart.domain.member.Member;
 import cart.domain.product.Product;
