+++ src/main/java/cart/domain/order/OrderItem.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.order;
 
 import cart.domain.product.Product;
