+++ src/main/java/cart/domain/point/BasicPointPolicy.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.point;
 
 import cart.domain.common.Money;
