+++ src/main/java/cart/dto/CartItemResponse.java
@@ -1,6 +1,6 @@
 package cart.dto;
 
-import cart.domain.CartItem;
+import cart.domain.cartItem.CartItem;
 
 public class CartItemResponse {
     private Long id;
