+++ src/test/java/cart/domain/MemberTest.java
@@ -2,6 +2,7 @@ package cart.domain;
 
 import cart.domain.member.Member;
 import cart.domain.member.Password;
+import cart.domain.point.Point;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
