+++ src/test/java/cart/domain/PointTest.java
@@ -1,5 +1,6 @@
 package cart.domain;
 
+import cart.domain.point.Point;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
