+++ src/test/java/cart/repository/OrderRepositoryTest.java
@@ -4,9 +4,9 @@ import cart.dao.MemberDao;
 import cart.dao.MemberEntity;
 import cart.dao.OrderDao;
 import cart.dao.ProductDao;
-import cart.domain.Order;
-import cart.domain.OrderItem;
 import cart.domain.member.Member;
+import cart.domain.order.Order;
+import cart.domain.order.OrderItem;
 import cart.domain.product.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
